# Include quantum excitation