--- a/include/trackcpp/accelerator.h
+++ b/include/trackcpp/accelerator.h
@@ -28,8 +28,7 @@ public:
   Accelerator(const double& energy=-1);
   double                  energy;              // [eV]
   bool                    cavity_on;
-  bool                    radiation_on;
-  bool                    quantdiff_on;
+  int                     radiation_on;
   bool                    vchamber_on;
   int                     harmonic_number;
   std::vector<Element>    lattice;