--- a/include/trackcpp/auxiliary.h
+++ b/include/trackcpp/auxiliary.h
@@ -168,6 +168,7 @@ int sgn(T val) {
 }
 
 
+void set_random_seed(unsigned rnd_seed);
 double gen_random_number();