--- a/include/trackcpp/passmethods.hpp
+++ b/include/trackcpp/passmethods.hpp
@@ -156,7 +156,8 @@ void bndthinkick(Pos<T>& pos, const double& length,
                  const std::vector<double>& polynom_b,
                  const double& irho,
                  const Accelerator& accelerator,
-                 const bool quantum_kick_on) {
+                 const bool quantum_kick_on,
+                 const double random_number=0) {
 
   T real_sum, imag_sum;
   calcpolykick<T>(pos, polynom_a, polynom_b, real_sum, imag_sum);
@@ -175,9 +176,9 @@ void bndthinkick(Pos<T>& pos, const double& length,
       radiation_constant*SQR(1+pos.de)*b2p*(1+irho*rx + (px*px+py*py)/2)*length;
 
     if (quantum_kick_on){
-      double p0 = accelerator.energy/light_speed;
-      double p0_SI = (accelerator.energy/light_speed) * electron_charge;
-      double gamma = accelerator.energy/M0C2;
+      const double p0 = accelerator.energy/light_speed;
+      const double p0_SI = (accelerator.energy/light_speed) * electron_charge;
+      const double gamma = accelerator.energy/M0C2;
       
       T dl_ds = (1 + rx*irho);
 
@@ -185,7 +186,7 @@ void bndthinkick(Pos<T>& pos, const double& length,
         * pow(abs(irho), 3)*pow(p0, 2)*p0_SI*dl_ds/pow(accelerator.energy, 2) *
         (length/KICK2);
 
-      T qkick = sqrt(d) * random_normal(0., 1.);
+      T qkick = sqrt(d) * random_number;
 
       pos.de += qkick;
     }
@@ -303,6 +304,12 @@ Status::type pm_bnd_mpole_symplectic4_pass(Pos<T> &pos, const Element &elem,
   double irho = elem.angle / elem.length;
   const std::vector<double> &polynom_a = elem.polynom_a;
   const std::vector<double> &polynom_b = elem.polynom_b;
+  //Necessary utils for quantum diffusion
+  double rn = 0;
+  std::random_device rand_dev;
+  std::mt19937  generator(rand_dev());
+  // std::mt19937  generator;
+  std::normal_distribution<double>  distr(0., 1.);
 
   global_2_local(pos, elem);
   edge_fringe(pos, irho, elem.angle_in, elem.fint_in, elem.gap);
@@ -310,7 +317,11 @@ Status::type pm_bnd_mpole_symplectic4_pass(Pos<T> &pos, const Element &elem,
     drift<T>(pos, l1);
     bndthinkick<T>(pos, k1, polynom_a, polynom_b, irho, accelerator, false);
     drift<T>(pos, l2);
-    bndthinkick<T>(pos, k2, polynom_a, polynom_b, irho, accelerator, accelerator.quantdiff_on);
+    if (accelerator.quantdiff_on){
+      rn = distr(generator);
+    }
+    bndthinkick<T>(pos, k2, polynom_a, polynom_b, irho, accelerator,
+    accelerator.quantdiff_on, rn);
     drift<T>(pos, l2);
     bndthinkick<T>(pos, k1, polynom_a, polynom_b, irho, accelerator, false);
     drift<T>(pos, l1);