--- a/include/trackcpp/passmethods.hpp
+++ b/include/trackcpp/passmethods.hpp
@@ -146,8 +146,7 @@ void strthinkick(Pos<T>& pos, const double& length,
 
     if (d_factor != 0) {
       // quantum excitation kick
-      T dl_ds = (1 + rx*sqrt(b2p));
-      T d = pow(sqrt(b2p), 3) * d_factor * dl_ds;
+      T d = pow(sqrt(b2p), 3) * d_factor;
       double random_number = gen_random_number();
       T qkick = sqrt(d) * random_number;
       pos.de += qkick;
@@ -188,7 +187,7 @@ void bndthinkick(Pos<T>& pos, const double& length,
     if (d_factor != 0) {
       // quantum excitation kick
       T dl_ds = (1 + rx*irho);
-      T d = d_factor * dl_ds;
+      T d = d_factor * pow(sqrt(b2p), 3) * dl_ds;
       double random_number = gen_random_number();
       T qkick = sqrt(d) * random_number;
       pos.de += qkick;
@@ -324,7 +323,7 @@ Status::type pm_bnd_mpole_symplectic4_pass(Pos<T> &pos, const Element &elem,
     const double p0_SI = p0 * electron_charge;
     const double gamma = accelerator.energy/M0C2;
     d_factor = CU * CER * reduced_planck_constant*pow(gamma, 4)/pow(electron_mass, 2)
-      * pow(abs(irho), 3)*pow(p0, 2)*p0_SI/pow(accelerator.energy, 2)*sl;
+      *pow(p0, 2)*p0_SI/pow(accelerator.energy, 2)*sl;
   }
 
   global_2_local(pos, elem);