… src/accelerator.cpp
… src/auxiliary.cpp
… src/commands.cpp
… src/diffusion_matrix.cpp
… src/flat_file.cpp
… src/output.cpp
… src/tests.cpp
… tests/si_v07_c05.txt
… tests/test-kickmap.cpp