@jquentino 2022-07-28
To keep the group members that were on vacation updated about this PR when they return:

In some talks with @fernandohds564, we decide to change the `radiation_on` property from a boolean to an enumerate, with the following options:

- 0 = no radiation,
- 1 = radiation damping,
- 2 = damping+quantum_excitation.

The idea is that the user can set one of the three options by passing the correspondent integer number or a string.
In this way, we can use `radiation_on` to control all available radiative effects without the need to set a new property.

The feature is being developed in the branch `modify-rad-properties` and will be merged into this PR when it is working.